Corrected fixed limit problem

+++ bit-lang/src/bit_spec.rs
@@ -17,8 +17,8 @@ use std::fmt;
 
 #[derive(Debug, PartialEq, Eq, Clone, Hash, PartialOrd, Ord)]
 pub enum BitRange {
-    Single(u8),
-    Range(u8, u8),
+    Single(u16),
+    Range(u16, u16),
     WholeWord,
     Literal(String),
 }
